Add notifications for error messages
Before, error messages were caught and printed to console. Now they're
more useful and are displayed above the input box in a Bootstrap Alert.
Try-catch blocks are moved one level up to make it easier to pass
errors to the Notification component, which now uses PropTypes to help
with validation.